+++ src/dom/lib/traverse-helpers.ts
@@ -1,19 +1,3 @@
-const polyfilledElementMatches = Element.prototype.matches
-	|| (Element.prototype as any).msMatchesSelector
-	|| (Element.prototype as any).webkitMatchesSelector;
-
-
-/**
- * Returns whether the given element matches the optional selector
- *
- * @internal
- */
-export function elementMatches<ElementType extends Element> (element: ElementType, selector: string|null = null) : boolean
-{
-	return (null === selector) || polyfilledElementMatches.call(element, selector);
-}
-
-
 /**
  * Fetches all siblings, in the given direction.
  * Will start from the given element, traverse in the given direction and fetch the first (or all) matches
@@ -23,7 +7,7 @@ export function elementMatches<ElementType extends Element> (element: ElementTyp
 export function fetchSiblings<ElementType extends HTMLElement> (
 	element: HTMLElement,
 	selector: string|null,
-	accessor: "previousElementSibling" | "nextElementSibling"
+	accessor: "previousElementSibling" | "nextElementSibling",
 ) : ElementType[]
 {
 	let sibling = element[accessor];
@@ -31,7 +15,7 @@ export function fetchSiblings<ElementType extends HTMLElement> (
 
 	while (sibling)
 	{
-		if (elementMatches(sibling, selector))
+		if (null === selector || sibling.matches(selector))
 		{
 			list.push(sibling as ElementType);
 		}