+++ src/lib/events.ts
@@ -2,15 +2,10 @@ import {EventName} from '../dom/events';
 
 /**
  * Creates a new custom event
+ *
+ * @deprecated Just use the CustomEvent constructor directly
  */
 export function createEvent (type: EventName, args: CustomEventInit): CustomEvent
 {
-	if (typeof CustomEvent !== "function")
-	{
-		const event = document.createEvent("CustomEvent");
-		event.initCustomEvent(type, args.bubbles, args.cancelable, args.detail);
-		return event;
-	}
-
 	return new CustomEvent(type, args);
 }