+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+2.2.1 (unreleased)
+=====
+
+* (improvement) Remove internal `elementMatches()` helper.
+* (improvement) Remove outdated `CustomEvent` workaround.
+* (deprecation) Deprecate `createEvent` helper.
+* (improvement) Remove outdated `MediaQueryList.addEventListener()` workaround.
+
+
 2.2.0
 =====